feat(base,ui)!: BzInputNumber<TValue> - generic over Blazor's numeric types, exact decimal math
BaseInputNumber and BzInputNumber take a TValue type parameter covering
the same numeric types Blazor's own InputNumber<TValue> supports - int,
long, short, float, double, decimal, bare or nullable - validated with
the same runtime check and message. TValue is inferred from the binding:
@bind-Value on an int field now works directly, no (double?) cast, no
transforming setter.

Per-type semantics follow Blazor's converter:
- an integral TValue rejects fractional input outright ("2.7" is
  invalid for an int field, not a 2)
- only a nullable TValue can rest empty; a non-nullable one keeps its
  last good value and reverts on blur, like a failed Blazor parse
- values clamp to the type's own range on top of Min/Max, so a step or
  a typed value can never overflow the conversion back

All internal math now runs in decimal (InputNumberValue<TValue> is the
one seam where TValue crosses to/from it), so integer and money math is
exact - three 0.1 steps make 0.3, not 0.30000000000000004. The cascaded
InputNumberContext carries decimal?, which keeps the Group / Input /
Increment / Decrement parts non-generic - only the root closed over a
type argument. Min/Max/Step/LargeStep stay double-typed parameters, so
existing markup literals (Step="0.5") are untouched; float/double
magnitudes beyond decimal's range saturate, snap overflow near the
decimal edge falls back to the clamped value.

BREAKING: code referencing the component type by name needs the type
argument (BzInputNumber<double?>); bindings are unaffected (TValue
infers). BzColorInput/BzColorOpacityInput pin TValue explicitly; docs
ApiFamilies moved to the open-generic typeof.

11 new type-behavior tests (int/short/long/float/decimal round-trips,
fraction rejection, nullable-empty contract, range pinning, unsupported-
type throw); the controlled-lag suite re-targeted to the generic root.

+++ src/Blaizio.Base/Components/InputNumber/InputNumberContext.cs
@@ -3,10 +3,12 @@ using Microsoft.AspNetCore.Components;
 namespace Blaizio;
 
 /// <summary>
-/// State a <see cref="BaseInputNumber"/> cascades to its group, input, and the increment / decrement
-/// buttons. The root owns the numeric value and ALL the math (parse, clamp to <see cref="Min"/> /
+/// State a <c>BaseInputNumber&lt;TValue&gt;</c> cascades to its group, input, and the increment /
+/// decrement buttons. The root owns the numeric value and ALL the math (parse, clamp to <see cref="Min"/> /
 /// <see cref="Max"/>, snap to the step grid) plus the press-and-hold repeat timer; the parts are thin -
-/// they read this context to render and call its semantic callbacks to change the value.
+/// they read this context to render and call its semantic callbacks to change the value. The value
+/// rides as <see cref="decimal"/> - the root's internal math type, whatever its <c>TValue</c> - which
+/// is what keeps this record, and every part reading it, non-generic.
 /// </summary>
 /// <param name="Value">The current numeric value, or <see langword="null"/> when the field is empty.</param>
 /// <param name="Min">The lowest allowed value, or <see langword="null"/> when unbounded below.</param>
@@ -30,9 +32,9 @@ namespace Blaizio;
 /// <param name="PressStart">A pointer pressed a stepper button (<c>+1</c> up, <c>-1</c> down): step once, then hold-to-repeat.</param>
 /// <param name="PressEnd">The pointer released / left the stepper button: stop repeating.</param>
 public sealed record InputNumberContext(
-    double? Value,
-    double? Min,
-    double? Max,
+    decimal? Value,
+    decimal? Min,
+    decimal? Max,
     bool CanIncrement,
     bool CanDecrement,
     bool Disabled,
